utils.cloudinit: split PhoneHomeServer logic
Currently, we can only start the PhoneHomeServer and wait for a phone call
in a single call/method. This small change it will allow clients of this
library to first start the server and wait for a phone call in a future
call.  It came from a QEMU ML discussion and probably they will not use
this change anytime soon, because of dependency policies but in any case
it is valid the change for other users.

Reference: https://www.mail-archive.com/qemu-devel@nongnu.org/msg871809.html
Signed-off-by: Beraldo Leal <bleal@redhat.com>