utils.network.ports refactoring
Currently we are assuming that UDP and TCP ports shares the the same domain,
which is not true. The same is valid for IPv4 vs IPv6. When asking for
free ports, makes sense to be specific about what kind of port we are
looking for. Because of that, going over all the possible domains to get
a free port is odd. Closes #5273

Signed-off-by: Beraldo Leal <bleal@redhat.com>